# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Organization Directory API",
        "description": "REST API for Organization Directory",
        "version": "1.0.0"
    },
    "paths": {
        "/api/buildings": {
            "get": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Вывод всех зданий с их организациями",
                "operationId": "3a9a966cf73ba9f32846a91c982a9459",
                "responses": {
                    "200": {
                        "description": "Вывод всех зданий с их организациями",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Building"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organizations/building/{building_id}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Список организаций в здании",
                "operationId": "779bab3c3dc50606bf8b6077cf3301de",
                "parameters": [
                    {
                        "name": "building_id",
                        "in": "path",
                        "description": "ID Здания",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций в здании",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Organization"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Building not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organizations/activity/{activity_id}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Список организаций по деятельности",
                "operationId": "b6440ac474a4c2262654a73d20012d58",
                "parameters": [
                    {
                        "name": "activity_id",
                        "in": "path",
                        "description": "ID Деятельности",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций по деятельности",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Organization"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Activity not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organizations/nearby": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Список организаций по географической близости",
                "operationId": "d95ab2d9ed79cabf5564449312232bbb",
                "parameters": [
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Широта",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "description": "Долгота",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Радиус в киллометрах (по умолчанию: 10)",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций по географической близости",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Organization"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organizations/area": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Список организаций в географической области",
                "operationId": "334db12989af0e152850afa9db0445d2",
                "parameters": [
                    {
                        "name": "lat1",
                        "in": "query",
                        "description": "Первая широта",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "lng1",
                        "in": "query",
                        "description": "Первая долгота",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "lat2",
                        "in": "query",
                        "description": "Вторая широта",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "lng2",
                        "in": "query",
                        "description": "Вторая долгота",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций в географической области",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Organization"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organizations/{id}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Сведения об организации",
                "operationId": "11826ee077ae441c61f2bee8476e2a14",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Organization ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сведения об организации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Organization"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Organization not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organizations/search": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Поиск организации",
                "operationId": "d6a844c687da3752fc5c80c956078c68",
                "parameters": [
                    {
                        "name": "activity",
                        "in": "query",
                        "description": "Поиск организаций по деятельности",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Поиск организаций по названию",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Результаты поиска организаций",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Organization"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Activity": {
                "title": "Activity",
                "description": "Activity model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Продовольствие"
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "parent": {
                        "$ref": "#/components/schemas/Activity"
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Activity"
                        }
                    },
                    "organizations": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Organization"
                        }
                    }
                },
                "type": "object"
            },
            "Building": {
                "title": "Building",
                "description": "Building model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "address": {
                        "type": "string",
                        "example": "Москва, улб. Ленина д. 1"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 55.7558
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 37.6176
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "organizations": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Organization"
                        }
                    }
                },
                "type": "object"
            },
            "Organization": {
                "title": "Organization",
                "description": "Organization model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "ООО Рога и Копыта"
                    },
                    "building_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "building": {
                        "$ref": "#/components/schemas/Building"
                    },
                    "phone_numbers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PhoneNumber"
                        }
                    },
                    "activities": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Activity"
                        }
                    }
                },
                "type": "object"
            },
            "PhoneNumber": {
                "title": "PhoneNumber",
                "description": "PhoneNumber model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "organization_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "number": {
                        "type": "string",
                        "example": "2-222-222"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "organization": {
                        "$ref": "#/components/schemas/Organization"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "API Key for authentication",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Buildings",
            "description": "Buildings"
        },
        {
            "name": "Organizations",
            "description": "Organizations"
        }
    ]
}